+++ frontend/packages/core/src/Feedback/stories/note-panel.stories.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import type { Meta } from "@storybook/react";
 
-import { ClipboardButton } from "../../button";
+import { ClipboardButton } from "../../Input";
 import { Link } from "../../link";
 import type { NotePanelProps } from "../note";
 import { Note, NotePanel } from "../note";
