+++ frontend/packages/core/src/Input/index.tsx
@@ -1,3 +1,4 @@
+export { Button, ButtonGroup, ClipboardButton, IconButton } from "./button";
 export { Checkbox, CheckboxPanel } from "./checkbox";
 export { default as DateTimePicker } from "./date-time";
 export { Form, FormRow } from "./form";
@@ -9,3 +10,4 @@ export { default as TextField } from "./text-field";
 export { default as ToggleButtonGroup, ToggleButton } from "./toggle-button-group";
 export type { ToggleButtonGroupProps } from "./toggle-button-group";
 export type { SelectOption } from "./select";
+export type { ButtonProps } from "./button";
