+++ frontend/packages/core/src/NPS/feedback.tsx
@@ -5,10 +5,9 @@ import { debounce } from "lodash";
 import { v4 as uuid } from "uuid";
 
 import { userId } from "../AppLayout/user";
-import { Button } from "../button";
 import { Alert } from "../Feedback";
 import type { SelectOption } from "../Input";
-import { Select, TextField } from "../Input";
+import { Button, Select, TextField } from "../Input";
 import { Grid } from "../Layout";
 import { client } from "../Network";
 import type { ClutchError } from "../Network/errors";
