+++ frontend/packages/core/src/Resolver/input.tsx
@@ -4,18 +4,18 @@ import type { clutch } from "@clutch-sh/api";
 import { DevTool } from "@hookform/devtools";
 import SearchIcon from "@mui/icons-material/Search";
 
+import { Alert } from "../Feedback";
+import { Button } from "../Input/button";
+import TextField from "../Input/text-field";
+import { useSearchParams } from "../navigation";
+import { client } from "../Network";
 import {
   Accordion,
   AccordionActions,
   AccordionDetails,
   AccordionDivider,
   AccordionProps,
-} from "../accordion";
-import { Button } from "../button";
-import { Alert } from "../Feedback";
-import TextField from "../Input/text-field";
-import { useSearchParams } from "../navigation";
-import { client } from "../Network";
+} from "../Surface/accordion";
 import { styled } from "../Utils";
 
 import type { ChangeEventTarget } from "./hydrator";
