+++ frontend/packages/core/src/Surface/card.tsx
@@ -12,9 +12,9 @@ import {
 import type { SpacingProps as MuiSpacingProps } from "@mui/system";
 import { spacing } from "@mui/system";
 
-import { IconButton } from "./button";
-import { styled } from "./Utils";
-import { Typography, TypographyProps } from "./typography";
+import { IconButton } from "../Input";
+import { Typography, TypographyProps } from "../typography";
+import { styled } from "../Utils";
 
 // TODO: seperate out the different card parts into various files
