+++ frontend/packages/core/src/Surface/index.tsx
@@ -1,0 +1,11 @@
+export {
+  Accordion,
+  AccordionActions,
+  AccordionDetails,
+  AccordionDivider,
+  AccordionGroup,
+} from "./accordion";
+export { Card, CardContent, CardHeader } from "./card";
+export { default as Paper } from "./paper";
+
+export type { CardHeaderSummaryProps } from "./card";
