+++ frontend/packages/core/src/Surface/paper.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import type { PaperProps as MuiPaperProps } from "@mui/material";
 import { Paper as MuiPaper } from "@mui/material";
 
-import { styled } from "./Utils";
+import { styled } from "../Utils";
 
 export interface PaperProps extends Pick<MuiPaperProps, "children"> {}
