+++ frontend/packages/core/src/index.tsx
@@ -1,15 +1,6 @@
-export {
-  Accordion,
-  AccordionActions,
-  AccordionDetails,
-  AccordionDivider,
-  AccordionGroup,
-} from "./accordion";
 export { userId } from "./AppLayout/user";
 export * from "./Assets/emojis";
 export * from "./Assets/icons";
-export { Button, ButtonGroup, ClipboardButton, IconButton } from "./button";
-export { Card, CardContent, CardHeader } from "./card";
 export * from "./Charts";
 export * from "./chip";
 export { default as Confirmation } from "./confirmation";
@@ -31,9 +22,9 @@ export {
 export { client } from "./Network";
 export * from "./NPS";
 export { default as ExpansionPanel } from "./panel";
-export { default as Paper } from "./paper";
 export { default as Resolver } from "./Resolver";
 export { Step, Stepper } from "./stepper";
+export * from "./Surface";
 export * from "./Utils";
 export { Tab, Tabs } from "./tab";
 export {
@@ -47,15 +38,12 @@ export {
   TreeTable,
 } from "./Table";
 export { default as Code } from "./text";
-export { default as TimeAgo } from "./timeago";
+export { default as TimeAgo } from "./Utils/timeago";
 export { Typography } from "./typography";
 export { default as ClutchApp } from "./AppProvider";
 export * from "./Layout";
 
 export type { BaseWorkflowProps, WorkflowConfiguration } from "./AppProvider/workflow";
-export type { ButtonProps } from "./button";
-export type { CardHeaderSummaryProps } from "./card";
-
 export type { WorkflowRemoveDataFn, WorkflowRetrieveDataFn, WorkflowStoreDataFn } from "./Contexts";
 export type { ClutchError } from "./Network/errors";
 export type { TypographyProps } from "./typography";
