+++ frontend/packages/core/src/stories/stepper.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import type { Meta } from "@storybook/react";
 
-import { Button, ButtonGroup } from "../button";
+import { Button, ButtonGroup } from "../Input";
 import type { StepperProps } from "../stepper";
 import { Step, Stepper } from "../stepper";
 import { styled } from "../Utils";
